// src/app/shared/models/CardToBeSubmit.ts
export interface CardToBeSubmit {
  cardindex?:    string;
  cardNo?:       string;
  bankName?:     string;
  friendlyName?: string;
  type?:         string;
  deleted?:    boolean;
  emiBankID?: string;
  emi_bank_id?: string;
  emiTenure?: string;
  emi_tenure?: string;
  emi?: boolean;
  is_emi?: boolean;
  number?: string;
  expiry?: string;
  cvv: string;
  name?: string;
  save?: Boolean;
  selected?: Boolean;
  card_index?: string;
  emiBank?: string;
  phone?: string;
  brand?: string;
}